Replace readdir with scandir
It should be a little faster since it avoids a bunch of syscalls
