+++ otherlibs/stdune-unstable/dune_filesystem_stubs/dune
@@ -1,7 +1,7 @@
 (library
  (name dune_filesystem_stubs)
  (public_name stdune-unstable.filesystem_stubs)
- (libraries)
+ (libraries unix)
  (foreign_stubs
   (language c)
-  (names readdir)))
+  (names scandir)))
