+++ otherlibs/stdune-unstable/dune_filesystem_stubs/dune_filesystem_stubs.ml
@@ -35,76 +35,63 @@ module File_kind = struct
   end
 end
 
-module Readdir_result = struct
-  [@@@warning "-37"]
-
-  (* The values are constructed on the C-side *)
-  type t =
-    | End_of_directory
-    | Entry of string * File_kind.Option.t
-end
-
-external readdir_with_kind_if_available_unix :
-  Unix.dir_handle -> Readdir_result.t = "caml__dune_filesystem_stubs__readdir"
-
-let readdir_with_kind_if_available_win32 : Unix.dir_handle -> Readdir_result.t =
- fun dir ->
-  (* Windows also gives us the information about file kind and it's discarded by
-     [readdir]. We could do better here, but the Windows code is more
-     complicated. (there's an additional OCaml abstraction layer) *)
-  match Unix.readdir dir with
-  | exception End_of_file -> Readdir_result.End_of_directory
-  | entry -> Entry (entry, UNKNOWN)
-
-let readdir_with_kind_if_available : Unix.dir_handle -> Readdir_result.t =
-  if Stdlib.Sys.win32 then
-    readdir_with_kind_if_available_win32
-  else
-    readdir_with_kind_if_available_unix
-
-let read_directory_with_kinds_exn dir_path =
-  let dir = Unix.opendir dir_path in
-  Fun.protect
-    ~finally:(fun () -> Unix.closedir dir)
-    (fun () ->
-      let rec loop acc =
-        match readdir_with_kind_if_available dir with
-        | Entry (("." | ".."), _) -> loop acc
-        | End_of_directory -> acc
-        | Entry (base, kind) ->
-          let k kind = loop ((base, kind) :: acc) in
-          let skip () = loop acc in
-          File_kind.Option.elim kind
-            ~none:(fun () ->
-              match Unix.lstat (Filename.concat dir_path base) with
-              | exception _ ->
-                (* File disappeared between readdir & lstat system calls. Handle
-                   as if readdir never told us about it *)
-                skip ()
-              | stat -> k stat.st_kind)
-            ~some:k
-      in
-      loop [])
-
 let catch_unix_error f x =
   match f x with
   | exception Unix.Unix_error (error, _, _) -> Error error
   | res -> Ok res
 
+external scandir_with_kind_if_available_unix :
+  string -> (string * File_kind.Option.t) list = "dune_scandir"
+
+let scandir_with_kind_if_available_win32 dir =
+  (* Windows also gives us the information about file kind and it's discarded by
+     [readdir]. We could do better here, but the Windows code is more
+     complicated. (there's an additional OCaml abstraction layer) *)
+  match catch_unix_error Unix.opendir dir with
+  | Error e -> Error e
+  | Ok dir ->
+    Fun.protect
+      ~finally:(fun () -> Unix.closedir dir)
+      (fun () ->
+        let rec loop acc =
+          match Unix.readdir dir with
+          | "."
+          | ".." ->
+            loop acc
+          | exception End_of_file -> List.rev acc
+          | entry -> loop ((entry, File_kind.Option.UNKNOWN) :: acc)
+        in
+        Ok (loop []))
+
+let scandir_with_kind_if_available :
+    string -> ((string * File_kind.Option.t) list, Unix.error) result =
+  if Stdlib.Sys.win32 then
+    scandir_with_kind_if_available_win32
+  else
+    fun dir ->
+  catch_unix_error scandir_with_kind_if_available_unix dir
+
 let read_directory_with_kinds dir_path =
-  catch_unix_error read_directory_with_kinds_exn dir_path
+  match scandir_with_kind_if_available dir_path with
+  | Error e -> Error e
+  | Ok entries ->
+    Ok
+      (entries
+      |> List.filter_map (fun (base, kind) ->
+             File_kind.Option.elim kind
+               ~none:(fun () ->
+                 match Unix.lstat (Filename.concat dir_path base) with
+                 | exception _ ->
+                   (* File disappeared between readdir & lstat system calls.
+                      Handle as if readdir never told us about it *)
+                   None
+                 | stat -> Some (base, stat.st_kind))
+               ~some:(fun kind -> Some (base, kind))))
 
-let read_directory_exn dir_path =
-  let dir = Unix.opendir dir_path in
-  Fun.protect
-    ~finally:(fun () -> Unix.closedir dir)
-    (fun () ->
-      let rec loop acc =
-        match readdir_with_kind_if_available dir with
-        | Entry (("." | ".."), _) -> loop acc
-        | End_of_directory -> acc
-        | Entry (base, _) -> loop (base :: acc)
-      in
-      loop [])
-
-let read_directory dir_path = catch_unix_error read_directory_exn dir_path
+(* XXX this function is subtly different from [read_directory_with_kinds]. it
+   does not skip [UNKNOWN] entries for which lstat failed. Is this
+   intentional? *)
+let read_directory dir_path =
+  match scandir_with_kind_if_available dir_path with
+  | Error e -> Error e
+  | Ok s -> Ok (List.map fst s)
