+++ test/expect-tests/dune_filesystem_stubs/dune
@@ -1,0 +1,14 @@
+(library
+ (name dune_filesystem_stubs_test)
+ (inline_tests
+  (deps
+   (sandbox always)))
+ (libraries
+  dune_filesystem_stubs
+  stdune
+  ppx_expect.config
+  ppx_expect.config_types
+  ppx_expect.common
+  base
+  ppx_inline_test.config)
+ (preprocess (pps ppx_expect)))
